Create Fork (takeover)
This commit marks the start of Ecconias fork of JavaLoader.
It will have a some features removed and others added.
It won't be compatible with the original JavaLoader, as some different design choices will be made.
The fork will be maintained in a way to please Ecconias usage experience, it is not meant for the broad public.

Things that changed:
- Now targets Java 17.
- Added Ecconia as author.
- Start with version 1.0.0.
- Target latest Velocity.
- License is now AGPLv3.

Things that have been removed:
- Example projects: Can be added in future again. I don't need them.
- Checkstyle: A different style will be used.
- Tests: Not relevant rn, can be re-added in future.
- License injection: Fork will not be provided as jar.
- Standalone version: Not needed by me.